Project_Chatbot/pyproject.toml
[project]
name = "project-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.17",
    "langchain-core>=1.3.2",
    "langchain-google-genai>=4.2.2",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "python-dotenv>=1.2.2",
]
